src/requirements.txt
# this file is generated with scripts/pipFreeze.sh, do not update in any other way
certifi==2017.4.17
chardet==3.0.4
click==6.7
idna==2.5
kafka-python==1.3.3
oauthlib==2.0.2
psutil==5.2.2
python-dotenv==0.6.4
pyzmq==16.0.2
requests==2.18.2
requests-oauthlib==0.8.0
TwitterAPI==2.4.5
urllib3==1.22
